add new methods to ReportRestController

+++ backend/src/main/java/reports/reports/repository/ReportRepository.java
@@ -3,6 +3,6 @@ package reports.reports.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import reports.reports.domain.Report;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
 
 }
